Move ProjectSelector to its own module
